requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.5
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
filelock==3.0.12
Markdown==3.3.4
mysql-connector-python==8.0.25
mysqlclient==2.0.3
Pillow==8.3.1
platformdirs==2.0.2
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.5.0
